check user permissions before deleting items

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -28,6 +28,7 @@ const Mutations = {
 
     return item;
   },
+
   updateItem(parent, args, ctx, info) {
     // first take a copy of the updates
     const updates = { ...args };
@@ -44,12 +45,17 @@ const Mutations = {
       info
     );
   },
+
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{ id title}`);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id }}`);
     // 2. Check if they own that item, or have the permissions
-    // TODO
+    const ownsItem = item.user.id === ctx.request.user.id;
+    const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission));
+    if(!ownsItem && !hasPermissions) {
+      throw new Error('You don\'t have permission to do that')
+    }
     // 3. Delete it!
     return ctx.db.mutation.deleteItem({ where }, info);
   },
